Add continuous scroll mode for song fields.
The default is now a continuous scroll, instead of the back-and-forth
bounce; this can be changed by setting `mode=b` in the field options.

It is also possible to change the padding from the default `'   '` with
the `padding="..."` option.

Example: `{song format="%(title)s",mode=c,padding=" | "}` will provide a
continuous scroll, separated with ` | `.

Note: Text will be stripped before padding.